requirements-dev.txt
# Все из requirements.txt
-r requirements.txt

# Development
ipython==8.15.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pytest==7.4.2
pytest-django==4.7.0
coverage==7.3.2

# Code quality
flake8==6.1.0
black==23.9.1
isort==5.12.0